1er requete JPQL

+++ src/main/java/tn/esprit/tp_foyer/entity/Chambre.java
@@ -17,7 +17,7 @@ public class Chambre implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idChambre;
+    private long idChambre;
 
     private int numeroChambre;
